Enhance CatScreen with image display and API error handling
This commit introduces several improvements to the CatScreen:
- Added Coil library to display cat images in `CatItem`.
- Implemented image loading states (loading, error) using `SubcomposeAsyncImage`.
- Updated `CatItem` to display more cat details (name, origin, length).
- Modified `CatScreen` to fetch cat list on initial composition using `LaunchedEffect`.
- Default search term in `CatScreen` changed to "a" for initial data display.
- Improved error handling in `CatViewModel` to log HTTP errors and display error messages.
- Replaced `Divider` with `HorizontalDivider` in `CatList` for Material 3 compliance.
- Updated `Cat` model: `minWeight` is now `Double?`.
- Added `.idea/sqlandroid.xml` to version control.

+++ app/src/main/java/com/example/jetpackcomposeretrofit/data/model/Cat.kt
@@ -25,7 +25,7 @@ data class Cat(
     @SerializedName("min_life_expectancy")
     val minLifeExpectancy: Int?,
     @SerializedName("min_weight")
-    val minWeight: Int?,
+    val minWeight: Double?,
     @SerializedName("name")
     val name: String?,
     @SerializedName("origin")
